--- src/main/java/Command.java
@@ -1,11 +1,0 @@
-enum Command {
-    LIST, MARK, UNMARK, DELETE, TODO, DEADLINE, EVENT, TIME, BYE, UNKNOWN;
-
-    public static Command fromString(String command) {
-        try {
-            return Command.valueOf(command.toUpperCase());
-        } catch (IllegalArgumentException e) {
-            return UNKNOWN;
-        }
-    }
-}
