+++ src/main/java/viktor/commands/Commandable.java
@@ -1,3 +1,7 @@
+package viktor.commands;
+
+import viktor.exceptions.ViktorException;
+
 public interface Commandable {
     void execute() throws ViktorException;
 }
