+++ src/main/java/viktor/commands/UnmarkCommand.java
@@ -1,10 +1,15 @@
+package viktor.commands;
 import java.io.IOException;
+import viktor.exceptions.ViktorException;
+import viktor.storage.Storage;
+import viktor.tasks.TaskList;
 
-public class DeleteTaskCommand implements Commandable {
+
+public class UnmarkCommand implements Commandable {
     private int taskNumber;
     private TaskList tasks;
 
-    public DeleteTaskCommand(int taskNumber, TaskList tasks) {
+    public UnmarkCommand(int taskNumber, TaskList tasks) {
         this.taskNumber = taskNumber;
         this.tasks = tasks;
     }
@@ -14,12 +19,11 @@ public class DeleteTaskCommand implements Commandable {
         if (taskNumber >= tasks.size()) {
             throw new ViktorException("You're asking for the impossible! That task doesn't exist.");
         }
-        System.out.println("\nI guess "+ tasks.getTask(taskNumber).getDescription() + " is no longer your concern.\n");
-        tasks.removeTask(taskNumber);
-        System.out.println('\n' + "Now you have " + tasks.size() + " remaining tasks.\n");
-
+        tasks.getTask(taskNumber).beUndone();
+        System.out.println("\n Oh you've yet to finish " + tasks.getTask(taskNumber).getDescription() 
+                + "? Don't forget: progress waits for no man\n");
         try {
-                Storage.save(tasks);
+            Storage.save(tasks);
         } catch (IOException e) {
             System.out.println("Ah something must've gone awry: " + e.getMessage() 
                     + " Well, mistakes are but a part of progress.");
