+++ src/main/java/viktor/exceptions/ViktorException.java
@@ -1,3 +1,4 @@
+package viktor.exceptions;
 public class ViktorException extends Exception {
     public ViktorException(String message) {
         super(message);
