+++ src/main/java/viktor/parser/DateParser.java
@@ -1,3 +1,4 @@
+package viktor.parser;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
