+++ src/main/java/viktor/storage/Storage.java
@@ -1,9 +1,16 @@
+package viktor.storage;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
+import viktor.tasks.Deadline;
+import viktor.tasks.Event;
+import viktor.tasks.Task;
+import viktor.tasks.TaskList;
+import viktor.tasks.Todo;
+
 
 public class Storage {
     private static final String FILE_PATH = "./ip/data/viktor.txt";
