+++ src/main/java/viktor/tasks/Deadline.java
@@ -1,3 +1,6 @@
+package viktor.tasks;
+
+import viktor.parser.DateParser;
 import java.time.LocalDateTime;
 import java.time.LocalDate;
