+++ src/main/java/viktor/tasks/Event.java
@@ -1,5 +1,7 @@
+package viktor.tasks;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import viktor.parser.DateParser;
 
 public class Event extends Task {
     private String type;
