+++ src/main/java/viktor/tasks/Task.java
@@ -1,3 +1,4 @@
+package viktor.tasks;
 public abstract class Task {
     protected String description;
     protected boolean isDone;
