+++ src/main/java/viktor/tasks/TaskType.java
@@ -1,3 +1,4 @@
+package viktor.tasks;
 public enum TaskType {
     TODO, DEADLINE, EVENT
 }
