+++ src/main/java/viktor/tasks/Todo.java
@@ -1,3 +1,4 @@
+package viktor.tasks;
 public class Todo extends Task {
     private String type;
