+++ src/main/java/viktor/ui/Viktor.java
@@ -1,3 +1,13 @@
+package viktor.ui;
+
+import viktor.commands.Commandable;
+import viktor.commands.ByeCommand;
+import viktor.tasks.TaskList;
+import viktor.tasks.Task;
+import viktor.exceptions.ViktorException;
+import viktor.parser.Parser;
+import viktor.storage.Storage;
+
 import java.io.IOException;
 import java.util.ArrayList;
